requirements.txt
google-cloud-storage==3.8.0
google-cloud-bigquery==3.27.0
google-cloud-logging>=3.0.0,<4.0.0
pytest==9.0.2
pytest-cov==7.0.0
coverage==7.13.2
